+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/modelconfig/SampleModelConfigLoader.scala
@@ -1,13 +1,15 @@
 package pl.touk.nussknacker.engine.management.sample.modelconfig
 
 import com.typesafe.config.{Config, ConfigValueFactory}
+import pl.touk.nussknacker.engine.component.ComponentExtractor
 import pl.touk.nussknacker.engine.modelconfig.{InputConfigDuringExecution, ModelConfigLoader}
 
 class SampleModelConfigLoader extends ModelConfigLoader {
 
   override def resolveInputConfigDuringExecution(inputConfig: Config, configWithDefaults: Config, classLoader: ClassLoader): InputConfigDuringExecution = {
+    val withExtractors = ComponentExtractor(classLoader).loadAdditionalConfig(inputConfig, configWithDefaults)
     InputConfigDuringExecution(
-      inputConfig
+      withExtractors
         .withValue("configLoadedMs", ConfigValueFactory.fromAnyRef(System.currentTimeMillis()))
         .withValue("duplicatedSignalsTopic", ConfigValueFactory.fromAnyRef(configWithDefaults.getString("signalsTopic")))
     )